+++ features/merchant/seller_menu_common/src/main/java/com/tokopedia/seller/menu/common/domain/usecase/GetUserShopInfoUseCase.kt
@@ -74,7 +74,7 @@ class GetUserShopInfoUseCase @Inject constructor(
 
         private const val SHOP_ID_KEY = "shopId"
 
-        fun createRequestParams(shopId: Int) = HashMap<String, Any>().apply {
+        fun createRequestParams(shopId: Long) = HashMap<String, Any>().apply {
             put(SHOP_ID_KEY, shopId)
         }
     }
