+++ features/merchant/seller_menu_common/src/main/java/com/tokopedia/seller/menu/common/domain/usecase/TopAdsDashboardDepositUseCase.kt
@@ -30,7 +30,7 @@ class TopAdsDashboardDepositUseCase @Inject constructor(private val graphqlRepos
 
         private const val SHOP_ID_KEY = "shopId"
 
-        fun createRequestParams(shopId: Int) = HashMap<String, Any>().apply {
+        fun createRequestParams(shopId: Long) = HashMap<String, Any>().apply {
             put(SHOP_ID_KEY, shopId)
         }
     }
